Update hangman.js
Removed the word object and used an API to fetch 28 random words from an online dictionary. These are sorted into the word object depending on their length, then a random number generator chooses from the desired difficulty level at random. The word is checked for non-alphabetical characters and then if any are present (or if none are present and this round has been won or lost), the word is discarded from the object. Once the word object's current level length is 1 or less, the API calls for a new batch of words.