Update hangman.js
Put everything inside an IIFE so as to hide the answer/wordObject from the user's console.
Changed the function which render the letters to only render them once the letter has been guessed - before, you could see in the HTML markup that there were spans with the correct letters already within.
Added a couple of functions to set and reset the display width of the hidden word. This ensures that the word is always fully-visible by scaling down when necessary.